fix(clippy): clean `cargo clippy --all-targets -D warnings` + fix flaky oauth test
- cargo clippy --fix mechanical lints (collapsible_if, redundant closures,
  useless format, manual split_once, etc.) across src/ + tests/.
- Targeted manual fixes: manual_clamp -> clamp, a match -> matches!,
  needless_range_loop / missing_transmute_annotations / unnecessary_literal_unwrap
  via local #[allow] where the lint is a false-positive or unsafe-FFI.
- Crate-level #![allow] in lib.rs for accepted design lints (result_large_err,
  too_many_arguments, type_complexity, large_enum_variant) + pedantic doc lints.
- #![allow(dead_code)] on shared test fixtures (tests/common, tests/webhooks_common)
  — each test binary uses only a subset of helpers.
- Serialize the three oauth::config tests that mutate process-global DRUST_OAUTH_*
  env vars (#[serial]); they raced under parallel `cargo test --lib` (pre-existing
  flake, not caused by the clippy pass).

+++ src/storage/garage.rs
@@ -196,7 +196,7 @@ impl GarageClient {
             let meta = item.context("garage list entry failed")?;
             out.push(ObjectSummary {
                 key: meta.location.to_string(),
-                size: meta.size as u64,
+                size: meta.size,
                 last_modified: meta.last_modified,
             });
         }
